✨ Chore: Move src folder to root

+++ pages/api/post/delete.tsx
@@ -1,19 +1,28 @@
 import { connectDB } from "@/util/database";
 import { ObjectId } from "mongodb";
 import { NextApiRequest,NextApiResponse } from 'next/types';
+import { getServerSession } from 'next-auth';
+import { authOptions } from '../auth/[...nextauth]';
 
 export default async function Handler(req:NextApiRequest , res: NextApiResponse) {
   if (req.method === "POST") {
-    try {
-      let db = (await connectDB).db("forum");
-      let result = await db
-        .collection("post")
-        .deleteOne({ _id: new ObjectId(req.body._id) });
-      res.status(200).json("제발되죠,,ㅡ,잔짜,,,,,나눈물나..")
-    } catch (error) {
-      res.status(500);
-    }
-  } else {
+    const session = await getServerSession(req,res,authOptions)
+    const db = (await connectDB).db("forum");
+    const foundOne = await db.collection('post').findOne({ _id: new ObjectId(req.body._id)})
+    if(foundOne.author === session.user.name){
+      try {
+        let db = (await connectDB).db("forum");
+        let result = await db
+          .collection("post")
+          .deleteOne({ _id: new ObjectId(req.body._id) });
+          res.status(200).json("Deleted Successfully")
+        } catch (error) {
+          res.status(500);
+        }
+      }else{
+        return res.status(500).json("Unmatch User");
+      }
+    } else {
     return res.status(500).json("Wrong request method");
   }
 }
